app/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
transformers
faiss-cpu
psycopg2-binary
nltk
requests
biopython
sacremoses
protobuf